add index switch and post gravity and fake like coefficient.

+++ app/Console/Commands/Test.php
@@ -59,7 +59,7 @@ class Test extends Command
             {
                 $postKey = 'post.'.$postId.'.data';
                 $after = $this->likeCount($postId);
-                $coefficient = intval(Redis::get('fake_like_coefficient'));
+                $coefficient = floatval(Redis::get('fake_like_coefficient'));
                 Redis::hmset($postKey , array('temp_like'=>fakeLike($after['like'] , $coefficient)));
             }
         }
